requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
celery==5.3.6
django-celery-beat==2.5.0
django-crontab==0.7.1
nltk==3.8.1
textblob==0.17.1
python-dateutil==2.8.2
gunicorn==21.2.0
